Move AIFramework.{h,cpp} into AI client directory
